requirements.txt
google-generativeai
pillow
cryptography
requests
